stackdocs/packages to track lsstsw package repos
The purpose of stackdocs/packages is to help find and get information
about the repositories cloned in lsstsw.

find_packages uses Eups to get all setup packages. It returns a dict of
Package instances which are representations of packages' repositories.

The code really does rely on being used in conjunction with lsstsw
because it doesn't know how to get the path to the cloned repository of
arbitrary packages cloned in conjunction with the stack.

The environment variable $LSSTSW_BUILD_DIR must be set for lsstsw
packages to be discoverable.

At the same time, it also relies on Eups to identify setup packages and
to get their setup versions.

For DM-4143.