Add jira, jirab and jirap reST roles
These reST roles make it easy to link to any JIRA ticket. The
jira_uri_template config variable sets the pattern for locating JIRA
ticket URLs. Default is to use the lsstcorp.org JIRA instance.

Use :jira: to in-text references. Multiple tickets can be concatenated
as well:

:jira:`DM-6181`
    -> DM-6181
:jira:`DM-6181,DM-6181`
    -> DM-6180 and DM-6181
:jira:`DM-6181,DM-6181,DM-6182`
    -> DM-6180,DM-6181, and DM-6182

The :jirap: and :jirab: roles do the same, but automatically enclose the
ticket(s) in parentheses or square brackets. The 'and' is also dropped
for compactness. These forms are useful in changelogs.

:jirap:`DM-6181`
    -> (DM-6181)
:jirap:`DM-6181,DM-6181`
    -> (DM-6180, DM-6181)
:jirap:`DM-6181,DM-6181,DM-6182`
    -> (DM-6180, DM-6181, DM-6182)