Obtain technote revision date and version from Git
This commit modifies the design doc/technote config system (ddconfig.py)
to obtain the version and last_revised metadata from the Git repo **if
those quantities are not specified in metadata.yaml already**. This
allows one to override automation for releases, while taking advantage
of Git's metadata for commit-by-commit work.

Uses GitPython to interact with Git (new dependency).

Includes mocked tests for read_git_branch and read_git_commit_timestamp
so that Git isn't a dependency of the unit tests.