Use TRAVIS_BRANCH env var for Git branch on Travis
Because of how Travis does shallow clones, it's been reported that the
existing method of using GitPython to infer the current Git branch
doesn't work. Instead we now detect when running on Travis CI and use
the available TRAVIS_BRANCH git branch instead.

Because we now act on the TRAVIS environment variable, I use monkeypatch
to temporarily unset the TRAVIS branch in tests that simulate a
non-travis environment.