replace ':' with '_' to avoid problem on windows
With epub output, a ':' was being put in the epub basename. On windows
this error in the filename (cannot have a ':' in windows filename),
caused building the epub output to die silently. I did this for the
three builders (epub, htmlhelp, and latex) that I found were replacing
space with '_'; I simply added a replacement of ':' with '_'.