Add find_package_docs
This uses the manifest.yaml file in a package's doc directory to discover
an eups package's "package" and "module" documentation directories.

This design is slightly different from DMTN-030 where we only
anticipated having directories for Python module documentation. But I'm
convinced that we need to add another tier of documentation for the eups
packages themselves. Some packages are data only, and some packages have
special installation instructions are are more apt to be discussed in
a package context than a python module/API context.

Adds test package to tests/data for unit tests.