Implement a Sphinx app runner function
run_sphinx is designed to emulate how Sphinx's own command line handler
works to create a Sphinx application object and handle exceptions from
it. Most of the exception handling code is taken from sphinx itself:

https://github.com/sphinx-doc/sphinx/blob/master/sphinx/cmdline.py

I've added a licenses.sphinx.txt to provide attribution for this code.

The Sphinx run is purposefully hard-coded. We can add configurability
later as the needs are recognized, but as an MVP, allowing for
configurability just increases the initial complexity.