Add get_project_content_commit_date function
This is a primitive function that tries to automatically determine the
revision date of a Sphinx documentation project based on the most recent
commit to content files in the Git repository. I've hard-coded it with
extensions for Sphinx content files (rst, png, etc.).

Note that gitpython sometimes has trouble finding commits for files in
sub-directories. That's why I'm working around the IOError for now.