Add edit_url to technote jinja context
This edit_url Jinja variable is built to be the GitHub edit link for
that specific technote's index.rst, given the computed branch.

The github_url is already provided through the metadata.yaml of each
technote, and the branch is computed from Git (but defaults to master).

Note I'm using the /blob/ rather than the /edit/ path so that an
external visitor doesn't immediately run into GitHub auth issues when
following that link.