Make object cross-references by parsing rst
This makes the field type and default into Python cross references. To
do this, I'm make reStructuredText that uses the ``:py:obj:``
cross-referencing role and then parsing that.

Parsing with state.nested_parse() initially presented problems but the
key was to wrap the content lines in a docutils.statemachine.ViewList
and provide that ViewList to both nested_parse and to
switch_source_input.

I've wrapped this rst parsing code into reusable methods that could be
further refactored out for reuse.