Create lsst-configs directive
This directive lets us present configuration fields for standalone
configs the same way that configuration fields are presented for Configs
associated with Tasks.

I'm not sure if this will be used in production, but it does allow me to
test presentations for pex.config Fields that aren't associated with
Task configuration classes.