Add format_registryfield_nodes for RegistryField
This docutils node generator creates documentation for
lsst.pex.config.RegistryField configuration fields. It's similar to
ConfigChoiceFields, but I needed to tweak it to show the configurable
class, rather than the configuration class associated with the
configurable.

I've also added RegistryFields as things not to show with general task
configurations (lsst-task-configs directive) but instead with the
lsst-subtasks directive.