Add lsst-config directive and role for cross-refs
This works similarly to the lsst-task directive and role, except it's
built for labelling topics about lsst.pex.config.Config classes.

We do this so that standalone Config topics have a reference target, and
so that the ConfigField, ConfigChoiceField, and ConfigDictField fields
can link to the topic pages for those configs.