Improved compactified config field presentation
Rather than presenting each key attribute of a config field as a
separate item in the definition list, this commit attempts to
consolidate some of that information into a slightly denser
presentation.

For example, the type that a Field accepts, and the type of the field
itself are combined.

Whether a field is optional is also included in the Field type
definition.

Whether a field is Multi-selection or Single-selection is also included
in the Field type.

The ordering of items in the definition list is also consistent. The
default is always first. Then the Field type. Then any additional,
unusual characteristics of the field that can't fit in the Field type
definition.