Create task topic marker directives with summaries
This commit takes the original task topic marker directives, and
substantially improves them. The goal here is to put summary
information into the environment so that when we list task topics we can
information like the summary, and what type of topic is it (a command
line task versus a task versus a configurable, etc.).

- Refactors task topic directive from crossrefs into the topics module
- New base class for all topic directives to keep the internal
  functionality the same (same schema for the env attribute)
- All topics now put data into the lsst_task_topics environment
  attribute.
- Store the summary of the task. The summary comes from
  either of two sources:

  1. The one-sentence summary of the Task class (default).
  2. The content of the lsst-task-topic directive.

  This is stored in the summary_node key in the lsst_tasks environment
  attribute. It'll be used to generate task listings or toctrees.