Create task topic listing / toctree directives
- Adds directives:

  - lsst-tasks
  - lsst-cmdlinetasks
  - lsst-pipelinetasks
  - lsst-configurables
  - lsst-configs

These can list topics by search along a Python namespace. The directives
can _also_ serve as a toctree for topic pages that appear in a
directory.

A big feature of these directives is that they show not only the class
name (the Task or Config's name), but also shows the summary describing
that task.