Support ConfigurableWrapper
If an item in an lsst.pex.config.Registry is wrapped in a
ConfigurableWrapper than it does not have __module__ and __name__
attributes like most tasks seem to. Instead, the task itself is
accessible from a _target attribute on the ConfigurableWrapper.

To handle the most generic case, the introduction can also fall back to
accessing the __class__ attribute and getting the __module__ and
__name__ attributes from that.