Get the eups tag from EUPS_TAG env variable
The EUPS_TAG environment variable is part of the Jenkins environment.
This way the "current release" information on the homepage of
https://pipelines.lsst.io will be dynamically based on EUPS_TAG, rather
than manually set from the version string in a pipelines_lsst_io conf.py
file. In other words, daily and weekly builds of pipelines.lsst.io will
be labeled as such.

There are now three default substitutions in the rst epilog:

- eups-tag (the EUPS_TAG)
- eups-tag-mono (the EUPS_TAG, but in monospace face)
- eups-tag-bold (the EUPS_TAG, but in bold face)

These substitutions replace the current-version substitution.

This also adds these configuration variables to the conf.py namespace:

- release_eups_tag: this is EUPS_TAG, or ``d_latest`` if not set.
- release_git_ref: this is the Git tag equivalent of EUPS_TAG, or
  "master" if it refers to "d_latest" or "w_latest".
- version: this is EUPS_TAG, or if a version string from a module's
  __version__ attribute was set, then it'll be that.
- release: same as "version"

This is a bit complicated at the moment and might be good to further
rationalize and document.