Add doxygen.conf file parsing
DoxygenConfiguration.from_doxygen_conf parses Doxygen configuration
files to create a new DoxygenConfiguration instances. This method is
simplified to only look for the tags we are interested in parsing from
the Doxygen configurations of individual packages:

- INPUT
- EXCLUDE
- EXCLUDE_PATTERNS
- EXCLUDE_SYMBOLS

This classmethod will be used to load the configs of individual
packages, and combine them into a singular monolithic doxygen
configuration.

Some of the parsing logic is guided by the makeDocs script in
lsst/lsstDoxygen.