Add preprocess_package_doxygen_conf function
This function plays the role of sconsUtils to process a doxygen.conf.in
file into a full configuration that includes INPUT paths for the include
directories.

The src, python, and example directories aren't included because they're
not needed to build a C++ API reference.

Added fake src and include directories to "package_alpha" to test
preprocess_package_doxygen_conf.