Migrate to setup.cfg setuptools configuration
As of setuptools 30.3.0 pretty much all of the package we used to set up
in setup.py's code can now be declared statically in setup.cfg.

The big different in practice is that now the "dev" extra no longer also
brings in the technote and pipelines extras; you need to install all of
them to set up a complete environment.