Add get_tag_entity_names function
This function uses the doxylink internals to extract API entity names
from the doxygen tag file. This will become the basis of future
functionality that lets us work with the doxygen tag file to introspect
the known C++ APIs.

To test this, I zipped and added a real doxygen tag file to the test
data. The tag_path test fixture unzips this data file during a test
session and makes it available to test functions.