Use a Doxygen defaults file
We discovered that the "base" package includes a doc/base.inc file that
provides a comprehensive set of defaults that can be overriden by more
specific configurations at build time. This commit attempts to capture
that same architecture, but now with all Doxygen configuration
consolidated into Documenteer (again, to descope lsstDoxygen,
sconsUtils, and base from the role of documentation builds).

- Embedded the base.inc file from lsst/base into Documenteer.
- Added an "include_path" attribute to DoxygenConfiguration that's
  handled specially at rendering to provide the @INCLUDE_PATH tag.
- Added a --doxygen-conf flag to stack-docs build to let a user choose
  an alternative configuration file (which might be useful for prototyping
  changes).