autocppapi creates fallback if no doxygen
If the doxylink symbol map isn't loaded, the autocppapi directive can
now fall back to displaying a message if C++ API information is
unavailable. Before, it would raise a RuntimeError and stop the build.

This change is important because single-package builds may not have a
Doxygen tag file available.

No longer need to conditionally add the lsstcc linking role; doxylink
will print the contents of the directive and not autocppapi will of
course fall back as well.