Support sphinxcontrib-bibtex 2.0+
sphinxcontrib-bibtex 2.0 requires the bibtex_bibfiles configuration,
which previously used to be part of the bibliography directive. This
commit enables that configuration for both old
(documenteer.sphinxconfig.technoteconf)
and new (documenteer.conf.technote) configuration APIs for technotes.

Also adds the bibtex_default_style config.

sphinxcontrib-bibtex is now constrainted to version >=2.0.0 because of
this breaking API change.