Organize docs for pydata theme headers
The nav bar in the pydata-sphinx-theme top header is populated from
**root level** toctree entries. This means we now need to be very
judicious about the toc tree organization to make the best use of this
theme feature.

- Moved the installation page to the homepage of the docs (will revise
to make it more compact)

- Make the change log a hidden toc tree item so it's only linked from
the nav bar

- Moved the API reference into the developer section and created a
homepage for the developer guide

- Shortened the page names of the section homepages for
sphinx-extensions/, pipelines/, and technotes/ to work better in the
homepage nav.