Vendor Rubin Style Dictionary 0.3 into CSS
This is a proof-of-concept / exploration of vendoring the CSS from Rubin
Style Dictionary (https://github.com/lsst-sqre/rubin-style-dictionary)

Overall more work needs to be done to get a complete set of design
tokens necessary to fully customize the pydata-sphinx-theme CSS; though
doing this is a great idea because it'll give us design consistency from
the RSP to the Sphinx docs, and in the future to technotes.

Also note that the vendoring is manually; potentially in the future we
should have a script that inserts the vendored code? Potentially even
use Sass? Most likely the vendoring updates should be done manually as a
dev process rather than have the version of Rubin Style Dictionary used
by Documenteer float.