Organize docs so project types are a top-level
Rather than creating a /project-guides/ top-level section, we're now
putting User guides, Pipelines, and Technotes into their own top-level
sections. This seems to make sense because each is treated separately
and we can then use captions within each section to segment the primary
sidebar.

The main difficulty with this is that it stretches the top-level nav
bar, forcing us to put some items in the "More" drop-down.