Add Rubin logo to technote
This adds the Rubin imagotypes (light and dark) as the logo for Rubin
technotes.

The logos are sourced from Rubin Style Dictionary, an installed npm
package, and are copied in the documenteer source using a webpack
plugin, copy-webpack-plugin. Assets sourced from Rubin Style Dictionary
reside in src/documenteer/assets/rsd-assets. With this change we can
potentially drop the other assets that are directly maintained in
documenteer so that all assets are derived from Rubin Style Dictionary.