Fix running help subcommand on stack/package-docs
The issue is that the project root discovery as being done in the main
command entrypoint before being passed to sub-commands like build and
clean. This meant that even "help" was running code to discover the root
of the project directory.

Now the raw root directory input is being passed to the commands that
need it, and those commands are running the package discovery.

Now package-docs build -h or package-docs help build will work.