Add githubbibcache Sphinx extension
The purpose of this extension is to let a user configure one or more
public GitHub repositories that contain bibtex files. The configured
bibtex files are automatically downloaded if not present and configured
for inclusion with the sphinxcontrib-bibtex extension.

We'll use this extension to download and cache bibtex files from
lsst/lsst-texmf. This replaces the need to commit these shared bibtex
files in individual repositories.