Upgrade to Sphinx 7 and require docutils>=0.20
It turns out that docutils 0.18 and 0.19 have a bug where the
bibliography directive from sphinxcontrib-bibtex extension add an
extraneous div element that throws off the document layout (noticed in
implementing bibliographies in technotes).

This bug is solved with docutils 0.20, which in turn requires us to use
Sphinx 7. Luckily it seems we can adopt Sphinx 7 without additional
changes.