Use Source Sans 3 variable font in technotes
The variable font (in .ttf.woff2 format) is obtained from
https://github.com/adobe-fonts/source-sans/tree/release/WOFF2/VF

These are version 3.052

Webpack takes responsibility for loading the webfont and then ensuring
that the CSS points to the right URL for it. In the course of doing
this, it also renames the font assets with a hash. To accomodate this,
I've added added extend_static_paths_with_asset_extension to the config
utilities functions. This helper function lets us glob the paths in
assets/ with a matching extension. This is needed because webpack
generates hashed filenames for the webfont assets.

Also, to ensure the relative paths work I've also changed to webpack
settings so that the rubin-technote.css is output to the root of
assets/, which is where webpack also deposits the font files.