Add interface to authordb.yaml file
The AuthorDb interface is based on one we wrote into
lsst-sqre/lsst-templatebot-aide, and adapted to use requests.

To improve on the latex encoding to unicode conversion, I'm adopting
pylatexenc. This is a terrific package that replaces a lot of work that
we'd used to to with pandoc or with fragile code of ours.