Use autodoc_pydantic and custom template
Originally from Gafaelfawr;
https://github.com/lsst-sqre/gafaelfawr/blob/main/docs/_templates/autosummary_core/pydantic_model.rst

This template ensures that we don't document inherited members from the
Pydantic BaseModel.

We benefit from autodoc_pydantic in two ways. First, it provides a
better presentation of Pydantic models.

Second, in conjunction with the custom automodapi template, we can avoid
a lot of API linking errors by omitting the base methods from BaseModel
in the API docs for a model.