Vendor parse_line_num_sec from Sphinx
remotecodeblock uses the same line highlighting syntax as the
LiteralIncludd directive. In Sphinx 7 this function was imported as
parselineno, but in Sphinx 8 it became parse_line_num_sec after a
migration period. To avoid this migration, and  since the function is
fairly self-contained, we're vendoring it directly into the Sphinx
extension here.